Correções no automapping
Correções do automapping e nos mapeamentos das classes. Ajustes da
interface de login.

+++ SisLar/Model/Repository/IRepositorio.cs
@@ -8,7 +8,7 @@ namespace SisLar.Model.Repository
 {
     public interface IRepositorio<T>
     {
-        T Retorna(int handle);
+        T Retorna(Int64 handle);
 
         IQueryable<T> RetornaTodos();
